+++ decision-tree-app/src/App.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import './index_backup.css';
+import './index.css';
 import { Routes, Route } from 'react-router-dom';
 import Header from './components/Header.jsx';
 import Home from './pages/Home.jsx';
