+++ decision-tree-app/src/components/Header.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function Header() {
+  return (
+    <header>
+      <h2>Decision Tree</h2>
+    </header>
+  );
+}
