+++ landing/README.md
@@ -1,0 +1,13 @@
+# Parsana Energy – Landing Page 🇮🇷⚡
+
+نمونهٔ اولیهٔ وب‌سایت تک‌صفحه‌ای برای معرفی خدمات «گروه مهندسی پارسانا انرژی».
+
+## راه‌‌اندازی سریع
+
+```bash
+git clone https://github.com/<YOUR-USERNAME>/parsana-energy-site.git
+cd parsana-energy-site
+# فایل‌ image_2025-07-13_16-17-37.png و تصاویر سرویس‌ها را
+# (battery.jpg, generator.jpg, solar.jpg) در پوشهٔ images/ قرار دهید
+open index.html   # یا live-server برای پیش‌نمایش
+```
