+++ landing/index.html
@@ -240,7 +240,9 @@
     </div>
   </footer>
 
-  <script type="module" src="/parsanaenergy/assets/index-Dmz9T8Sa.js"></script>
-  <script src="js/main.js"></script>
+  <!-- <script type="module" src="/parsanaenergy/assets/index-Dmz9T8Sa.js"></script> -->
+  <!-- <script src="js/main.js"></script> -->
+  <div id="root"></div>
+  <script type="module" src="/src/main.jsx"></script>
 </body>
 </html>
