+++ landing/src/App.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import './index.css';
+
+export default function App() {
+  return (
+    <div className="app">
+      <h1>Parsana Energy Landing</h1>
+    </div>
+  );
+}
