+++ landing/src/main.jsx
@@ -1,13 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { BrowserRouter } from 'react-router-dom';
 import App from './App.jsx';
-import './index.css';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
-    <BrowserRouter basename="/parsanaenergy/">
-      <App />
-    </BrowserRouter>
-  </React.StrictMode>,
+    <App />
+  </React.StrictMode>
 );
